+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed Resizer to ChapterResizer
 - Renamed SearchTextBox to ChapterSearchTextBox
 - Renamed SpacingStackPanel to ChapterStackPanel
+- Renamed SplitButton to ChapterSplitButton
 ### Supported .Net Versions
 - .Net Core 3.0
 - .Net Framework 4.5